Move LinearMemory trait and VMMemoryDefinition to wasmer_vm (from wasmer_types)
