+++ lib/api/src/sys/externals/memory_view.rs
@@ -4,7 +4,8 @@ use std::convert::TryInto;
 use std::marker::PhantomData;
 use std::mem::MaybeUninit;
 use std::slice;
-use wasmer_types::{LinearMemory, Pages};
+use wasmer_types::Pages;
+use wasmer_vm::LinearMemory;
 
 use super::memory::MemoryBuffer;
 use super::Memory;
