+++ lib/vm/src/lib.rs
@@ -45,7 +45,7 @@ pub use crate::function_env::VMFunctionEnvironment;
 pub use crate::global::*;
 pub use crate::imports::Imports;
 pub use crate::instance::{InstanceAllocator, InstanceHandle};
-pub use crate::memory::VMMemory;
+pub use crate::memory::{LinearMemory, VMMemory, VMMemoryDefinition};
 pub use crate::mmap::Mmap;
 pub use crate::probestack::PROBESTACK;
 pub use crate::sig_registry::SignatureRegistry;
@@ -61,9 +61,9 @@ pub use crate::vmcontext::{
 };
 pub use wasmer_types::LibCall;
 pub use wasmer_types::MemoryError;
+pub use wasmer_types::MemoryStyle;
 use wasmer_types::RawValue;
 pub use wasmer_types::TableStyle;
-pub use wasmer_types::{MemoryStyle, VMMemoryDefinition};
 pub use wasmer_types::{TargetSharedSignatureIndex, VMBuiltinFunctionIndex, VMOffsets};
 
 #[deprecated(
